refactor(suite): optimize how often fees are fetched in send form
!!! WORK IN PROGRESS !!! do not merge yet

This commit introduces a new approach to fetching fees.

Previous state:
Fees for every wallet account were fetched periodically after 10 blocks. This was problematic for various reasons:
1. The data weren't often needed.
2. Useless network requests.
3. Some fees were outdated, because fees aren't connected to blocks in any way - in fact, fee can update multiple times in between two blocks!

New state:
This solution mainly builds on the idea that fees aren't connected to blocks. This link is now removed. There is a dedicated `useFees` hook that is supposed to be called for every route that needs to display fees. The hook is responsible for decision on how often the fee should be fetched (decided manually in `networkToInterval` function). That way all logic behind fee fetching is located in one file. Consequence of this decision is that `updateFeeInfoThunk` is now "dumb" - it fetches new fees whenever it is called. While this solution is more verbose (`useFees` has to be called everywhere fees are needed), I think its explicit relationship improves readability of the code.

What needs to be done before this commit is mergeable:
1. Set network refresh intervals in `networkToInterval` function.
2. Call `useFees` for every route that needs fees (swap, sell, etc.).
3. Move `useFees` hook to better location so that it can be used in `suite-native/module-send/src/screens/SendOutputsScreen.tsx:174`, too.

Relates to #15087